+++ changelog/changelog.md
@@ -10,6 +10,10 @@
 
 - version list could exceed page count
 
+### Added
+
+- more pp stats in scorestats (requires `-all` flag)
+
 ## [3.18.0] - 2024-07-21
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot/commit/aede1ee3ca4a486312fd351b4d7c672a689d2188)